Game1
This is one of the first games I made for a class. We were expected to make a PERL program that used if-else statements but I was having fun and decided to make a short text based games with some concepts I was learning about about in my spare time and thought would make the game more interesting.