Football Game
I was working on a football game that would have a gui and take real stats from team websites and apply those to the playable characters. This was a bit ambitious for me early on and as I lost interest I now can't remember where I left off and where my files are. FBG is the main file that uses footballGameMenu, and FootballGame1. PygameAttempt3 is a collection of files from when I tried to make the playable football part of the game. Selecting MainQuiz in the zip file will show the code I wrote in order to make the beginnings of a football game with animated sprites, controllable movement and opponents with simple instructions.